+++ test/gigagrad-emnist.cpp
@@ -24,7 +24,7 @@ enum class DataType : uint8_t
 
 size_t SizeOf(DataType dtype)
 {
-    switch(dtype)
+    switch (dtype)
     {
     case DataType::U8:
         return sizeof(uint8_t);
@@ -53,23 +53,23 @@ struct ParsedDataFile
 ParsedDataFile LoadDataFile(const char *filename)
 {
     FILE *file = fopen(filename, "rb");
-    if(!file)
+    if (!file)
     {
         fprintf(stderr, "File %s not found\n", filename);
         exit(1);
     }
     uint32_t magic = 0;
-    if(fseek(file, 0, SEEK_SET))
+    if (fseek(file, 0, SEEK_SET))
     {
         fprintf(stderr, "Failed to seek: %d\n", ferror(file));
         exit(1);
     }
-    if(fread(&magic, sizeof(uint32_t), 1, file) < 1)
+    if (fread(&magic, sizeof(uint32_t), 1, file) < 1)
     {
         fprintf(stderr, "Failed to read from file: %d\n", ferror(file));
         exit(1);
     }
-    if(static_cast<uint16_t>(magic & 0xFFFF) != 0)
+    if (static_cast<uint16_t>(magic & 0xFFFF) != 0)
     {
         fprintf(stderr, "Invalid magic bytes at beginning of file %s (0x%x), exiting\n", filename, magic);
         exit(1);
@@ -77,7 +77,7 @@ ParsedDataFile LoadDataFile(const char *filename)
 
     uint8_t raw_dtype = (magic >> 16) & 0xFF;
     DataType dtype;
-    switch(raw_dtype)
+    switch (raw_dtype)
     {
     case (uint8_t)DataType::U8:
         dtype = DataType::U8;
@@ -103,21 +103,20 @@ ParsedDataFile LoadDataFile(const char *filename)
         __builtin_unreachable();
     }
 
-
     uint8_t ndim = (magic >> 24) & 0xFF;
 
     std::vector<size_t> shape(ndim);
     size_t offset = 4;
     size_t total_dataset_size = 1;
-    for(size_t i = 0; i < ndim; i++)
+    for (size_t i = 0; i < ndim; i++)
     {
         uint32_t dim = 0;
-        if(fseek(file, offset, SEEK_SET))
+        if (fseek(file, offset, SEEK_SET))
         {
             fprintf(stderr, "Failed to seek to offset %zu\n", offset);
             exit(1);
         }
-        if(fread(&dim, sizeof(dim), 1, file) < 1)
+        if (fread(&dim, sizeof(dim), 1, file) < 1)
         {
             fprintf(stderr, "Failed to read\n");
             exit(1);
@@ -128,36 +127,36 @@ ParsedDataFile LoadDataFile(const char *filename)
     }
 
     std::vector<uint8_t> data(total_dataset_size * SizeOf(dtype), 0);
-    if(fseek(file, offset, SEEK_SET))
+    if (fseek(file, offset, SEEK_SET))
     {
         fprintf(stderr, "Failed to seek to offset %zu\n", offset);
         exit(1);
     }
-    if(fread(data.data(), SizeOf(dtype), total_dataset_size, file) < total_dataset_size)
+    if (fread(data.data(), SizeOf(dtype), total_dataset_size, file) < total_dataset_size)
     {
         fprintf(stderr, "Failed to read dataset\n");
         exit(1);
     }
 
-    return { dtype, std::move(shape), std::move(data) };
+    return {dtype, std::move(shape), std::move(data)};
 }
 
 template <typename T>
-std::vector<float> CastToFloatAndNormalize(const std::vector<uint8_t> &input)
+std::vector<double> CastToFloatAndNormalize(const std::vector<uint8_t> &input)
 {
     size_t num_elements = input.size() / sizeof(T);
-    std::vector<float> result(num_elements);
+    std::vector<double> result(num_elements);
     const T *input_data = reinterpret_cast<const T *>(input.data());
-    for(size_t i = 0; i < num_elements; i++)
+    for (size_t i = 0; i < num_elements; i++)
     {
-        result[i] = static_cast<float>(input_data[i]) / 255.0;
+        result[i] = static_cast<double>(input_data[i]) / 255.0;
     }
     return result;
 }
 
-std::vector<float> CastToFloatAndNormalize(DataType dtype, const std::vector<uint8_t> &input)
+std::vector<double> CastToFloatAndNormalize(DataType dtype, const std::vector<uint8_t> &input)
 {
-    switch(dtype)
+    switch (dtype)
     {
     case DataType::U8:
         return CastToFloatAndNormalize<uint8_t>(input);
@@ -177,22 +176,22 @@ std::vector<float> CastToFloatAndNormalize(DataType dtype, const std::vector<uin
 }
 
 template <typename T>
-std::vector<float> ToOneHot(const std::vector<uint8_t> &input)
+std::vector<double> ToOneHot(const std::vector<uint8_t> &input)
 {
     size_t num_elements = input.size() / sizeof(T);
     const T *input_data = reinterpret_cast<const T *>(input.data());
     T max_val = 0;
-    for(size_t i = 0; i < num_elements; i++)
+    for (size_t i = 0; i < num_elements; i++)
     {
-        if(input_data[i] < 0)
+        if (input_data[i] < 0)
         {
             fprintf(stderr, "Tried to one-hot encode invalid value: %d\n", (int)input_data[i]);
             exit(1);
         }
         max_val = std::max(max_val, input_data[i]);
     }
-    std::vector<float> result(max_val * num_elements, 0.0f);
-    for(size_t i = 0; i < num_elements; i++)
+    std::vector<double> result(max_val * num_elements, 0.0f);
+    for (size_t i = 0; i < num_elements; i++)
     {
         T cur_val = input_data[i];
         result[i * max_val + cur_val] = 1.0f;
@@ -200,9 +199,9 @@ std::vector<float> ToOneHot(const std::vector<uint8_t> &input)
     return result;
 }
 
-std::vector<float> ToOneHot(DataType dtype, const std::vector<uint8_t> &input)
+std::vector<double> ToOneHot(DataType dtype, const std::vector<uint8_t> &input)
 {
-    switch(dtype)
+    switch (dtype)
     {
     case DataType::U8:
         return ToOneHot<uint8_t>(input);
@@ -221,8 +220,8 @@ std::vector<float> ToOneHot(DataType dtype, const std::vector<uint8_t> &input)
 struct Dataset
 {
     std::vector<size_t> shape;
-    std::vector<float> inputs;
-    std::vector<float> labels;
+    std::vector<double> inputs;
+    std::vector<double> labels;
 };
 
 Dataset LoadDataset(const char *directory, const char *dataset)
@@ -231,22 +230,22 @@ Dataset LoadDataset(const char *directory, const char *dataset)
     std::string label_name = std::string(directory) + "/emnist-mnist-" + dataset + "-labels-idx1-ubyte";
     ParsedDataFile images = LoadDataFile(image_name.c_str());
     ParsedDataFile labels = LoadDataFile(label_name.c_str());
-    std::vector<float> images_float = CastToFloatAndNormalize(images.dtype, images.data);
-    std::vector<float> labels_onehot = ToOneHot(labels.dtype, labels.data);
-    return { std::move(images.shape), std::move(images_float), std::move(labels_onehot) };
+    std::vector<double> images_float = CastToFloatAndNormalize(images.dtype, images.data);
+    std::vector<double> labels_onehot = ToOneHot(labels.dtype, labels.data);
+    return {std::move(images.shape), std::move(images_float), std::move(labels_onehot)};
 }
 
-void InitializeWeights(float *weight, size_t size_elts)
+void InitializeWeights(double *weight, size_t size_elts)
 {
     std::default_random_engine gen(0);
-    std::uniform_real_distribution<float> dist(-0.1f, 0.1f);
-    for(size_t i = 0; i < size_elts; i++)
+    std::uniform_real_distribution<double> dist(-0.1f, 0.1f);
+    for (size_t i = 0; i < size_elts; i++)
         weight[i] = dist(gen);
 }
 
 int main(int argc, const char **argv)
 {
-    if(argc != 2)
+    if (argc != 2)
     {
         fprintf(stderr, "Please specify exactly one argument: the directory of the EMNIST dataset\n");
         exit(1);
@@ -257,32 +256,32 @@ int main(int argc, const char **argv)
     constexpr size_t HiddenLayerSize = 40;
 
     gg::nn::Module network;
-    auto x = network.AddInput({ BatchSize, 28 * 28, 1 });
-    auto w1 = network.AddWeight({ HiddenLayerSize, 28 * 28 });
-    auto b1 = network.AddWeight({ HiddenLayerSize, 1 });
+    auto x = network.AddInput({BatchSize, 28 * 28, 1});
+    auto w1 = network.AddWeight({HiddenLayerSize, 28 * 28});
+    auto b1 = network.AddWeight({HiddenLayerSize, 1});
 
     auto x_bm = x.batchnorm();
     auto z1 = (w1 % x_bm) + b1;
     auto a2 = z1.relu();
-    auto w2 = network.AddWeight({ 10, HiddenLayerSize });
-    auto b2 = network.AddWeight({ 10, 1 });
+    auto w2 = network.AddWeight({10, HiddenLayerSize});
+    auto b2 = network.AddWeight({10, 1});
     auto z2 = (w2 % a2) + b2;
     auto result = z2.softmax(-2);
     gg::TrainingContext ctx = gg::CompileTrainingGraph<gg::codegen::BackendScalarC>(network, result, 0.005f);
 
-    w1.data() = new float[HiddenLayerSize * 28 * 28];
-    b1.data() = new float[HiddenLayerSize * 1];
-    w2.data() = new float[10 * HiddenLayerSize];
-    b2.data() = new float[10 * 1];
+    w1.data() = new double[HiddenLayerSize * 28 * 28];
+    b1.data() = new double[HiddenLayerSize * 1];
+    w2.data() = new double[10 * HiddenLayerSize];
+    b2.data() = new double[10 * 1];
     InitializeWeights(w1.data(), HiddenLayerSize * 28 * 28);
     InitializeWeights(b1.data(), HiddenLayerSize * 1);
     InitializeWeights(w2.data(), 10 * HiddenLayerSize);
     InitializeWeights(b2.data(), 10 * 1);
 
     size_t num_batches = train.shape[0] / BatchSize;
-    for(size_t iepoch = 0; iepoch < 100; iepoch++)
+    for (size_t iepoch = 0; iepoch < 100; iepoch++)
     {
-        for(size_t ibatch = 0; ibatch < num_batches; ibatch++)
+        for (size_t ibatch = 0; ibatch < num_batches; ibatch++)
         {
             x.data() = &train.inputs[BatchSize * 28 * 28 * ibatch];
             ctx.training_example = &train.labels[BatchSize * 10 * ibatch];
